refactor: 디렉토리 구조 변경
shared 폴더를 없애고 모두 utils로 이동
